No-BOM gate on job card creation; backfill ORD-091-26's BOM
Audit finding: 11 order items since mid-July entered production with zero
order_item_inventory lines — every deduction mechanism (stage-timed, QC
partial, dispatch settle) iterates those lines, so these builds consumed
parts that never left stock.

- Gate: creating a job card (normal or FG) for an item with no BOM lines
  now returns NO_BOM. The owner can consciously override — both upload
  modals offer the bypass on that error; other roles are told to get
  design to attach the BOM first.
- ORD-091-26 (99-pc FG order, already dispatched): BOM copied from the
  same drawing's ORD-093-26 item, scaled 20→99 pcs (pins 248+248, nuts
  495, bushes 495), deducted in the same stroke since the card is gone,
  item marked settled. Verified idempotent across two initDB runs.

The remaining 10 no-BOM items have no earlier order of the same drawing
to copy from — design attaches those by hand; the gate stops new ones.

+++ server/src/db/index.js
@@ -1828,7 +1828,7 @@ async function initDB(retries = 20, delayMs = 10000) {
           for (const [invId, qty] of lines) {
             await pool.query(
               `INSERT INTO order_item_inventory (order_item_id, inventory_item_id, qty, qty_deducted)
-               VALUES (176, $1, $2, $2)`, [invId, qty]);
+               VALUES (176, $1, $2::numeric, $2::numeric)`, [invId, qty]);
             await pool.query(
               'UPDATE inventory_items SET current_stock = current_stock - $1 WHERE id=$2',
               [qty, invId]);
